python_api/requirements.txt
# Recommended: Use Python 3.12.x for best package compatibility
# Python 3.14 is very new and many packages lack pre-built wheels

# Core Framework
fastapi==0.120.1
uvicorn[standard]==0.38.0

# Production WSGI Server
gunicorn==23.0.0

# Database
tortoise-orm==0.25.1
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Additional
python-multipart==0.0.20
pydantic==2.12.3
pydantic-settings==2.11.0
email-validator==2.3.0

# Testing (Development only)
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1

# Monitoring & Observability (Optional - uncomment for production)
# prometheus-client==0.21.0
# python-json-logger==2.0.7
